test: HubRoute 검색 결과 Pagination 테스트 (#19)
